Define TRPC inputs as separate schemas
