+++ .changeset/sharp-fans-exist.md
@@ -1,0 +1,5 @@
+---
+'@tradeb0t/core': patch
+---
+
+Define TRPC inputs as separate schemas
