+++ packages/core/src/api/trpc/schemas/AlgorithmName.ts
@@ -1,0 +1,3 @@
+import { z } from 'zod'
+
+export const ZAlgorithmName = z.string()
