+++ packages/core/src/api/trpc/schemas/GetOrdersOptions.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod'
 import { EOperationType } from 'src/domain'
 
-export const getOrdersOptions = z.object({
+export const ZGetOrdersOptions = z.object({
   from: z.date().optional(),
   to: z.date().optional(),
   securityTicker: z.string().optional(),
@@ -9,7 +9,7 @@ export const getOrdersOptions = z.object({
   runId: z.number().int().min(0).optional()
 })
 
-type ZGetOrdersOptions = z.infer<typeof getOrdersOptions>
+type ZGetOrdersOptions = z.infer<typeof ZGetOrdersOptions>
 
 // Interface looks better in UML diagrams
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
