+++ packages/core/src/api/trpc/schemas/Inputs.ts
@@ -1,0 +1,4 @@
+import { z } from 'zod'
+import { ZCreateOrderOptions } from './CreateOrderOptions'
+
+export const ZInputs = z.record(z.date().or(z.string()).or(z.number()).or(ZCreateOrderOptions))
