+++ packages/core/src/api/trpc/schemas/RunId.ts
@@ -1,0 +1,3 @@
+import { z } from 'zod'
+
+export const ZRunId = z.number().int().min(0)
