+++ packages/core/src/api/trpc/schemas/index.ts
@@ -1,1 +1,5 @@
 export * from './GetOrdersOptions'
+export * from './AlgorithmName'
+export * from './CreateOrderOptions'
+export * from './Inputs'
+export * from './RunId'
