--- packages/core/src/domain/models/connector/CreateOrderOptions.ts
@@ -1,8 +1,0 @@
-import { OperationType } from 'src/domain'
-
-export type CreateOrderOptions = {
-  operation: OperationType
-  ticker: string
-  lots: number
-  price: number
-}
