--- packages/core/src/domain/models/connector/index.ts
@@ -1,1 +1,0 @@
-export * from './CreateOrderOptions'
