+++ packages/core/src/domain/models/index.ts
@@ -1,3 +1,2 @@
-export * from './connector'
 export * from './persistent'
 export * from './memory'
